requirements.txt
django-crispy-forms==1.7.2
django-datetime-widget==0.9.3
django-markdown-deux==1.0.5
django-pagedown==1.0.4
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
Markdown==2.6.11
markdown2==2.3.5
mccabe==0.6.1
pycodestyle==2.4.0
six==1.11.0
urllib3==1.22
wrapt==1.10.11